% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsplyr.R
\name{include_jsplyr}
\alias{include_jsplyr}
\title{Link JS code.}
\usage{
include_jsplyr()
}
\value{
An \code{htmlDependency} object.
}
\description{
Include the \code{jsplyr} JavaScript code in a Shiny app. To be put
in the \code{UI} part of the app.
}
