% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-lazy-json.R
\name{tbl_lazy_json}
\alias{tbl_lazy_json}
\title{Create a lazy JSON tbl}
\usage{
tbl_lazy_json(session, json_name, compute_steps = list())
}
\arguments{
\item{session}{A shiny \code{session} object.}

\item{json_name}{A character.}

\item{compute_steps}{A list of compute steps to be
triggered when \code{compute()} is called.}
}
\value{
An object of class \code{tbl_lazy_json}: a list holding the shiny
\code{session}, a \code{state_id} keying the data in the browser, and the list of
\code{compute_steps} to run when the pipeline is computed.
}
\description{
Create a lazy JSON tbl
}
