% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-collect.R
\name{collect}
\alias{collect}
\alias{collect.tbl_lazy_json}
\title{Retrieve \code{JSON} data from the browser.}
\usage{
\method{collect}{tbl_lazy_json}(x, ..., raw = FALSE)
}
\arguments{
\item{x}{A \code{tbl_lazy_json} object.}

\item{...}{Unused. Provided for consistency with generic.}

\item{raw}{A logical, if \code{TRUE} returns \code{JSON} as character.}
}
\description{
Retrieve \code{JSON} data from the browser.
}
