% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-compute.R
\name{compute}
\alias{compute}
\alias{compute.tbl_lazy_json}
\title{Compute \code{JSON} data in the browser.}
\usage{
\method{compute}{tbl_lazy_json}(x, ...)
}
\arguments{
\item{x}{A \code{tbl_lazy_json} object.}

\item{...}{Unused. Provided for consistency with generic.}
}
\description{
Compute \code{JSON} data in the browser.
}
