% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-copy_to.R
\name{copy_to.ShinySession}
\alias{copy_to.ShinySession}
\title{Copy a local or remote data frame to the browser}
\usage{
\method{copy_to}{ShinySession}(dest, df, ...)
}
\arguments{
\item{dest}{A shiny \code{session} object.}

\item{df}{A local \code{data.frame} or a name of the
JSON data in the browser.}

\item{...}{Unused. Provided for consistency with generic.}
}
\description{
Copy a local or remote data frame to the browser
}
