% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb-distinct.R
\name{distinct}
\alias{distinct}
\alias{distinct.tbl_lazy_json}
\title{Keep distinct records of \code{JSON} data.}
\usage{
\method{distinct}{tbl_lazy_json}(.data, ...)
}
\arguments{
\item{.data}{A \code{tbl_lazy_json} object.}

\item{...}{Column names to determine uniqueness.
If empty, all columns are used.}
}
\description{
Keep distinct records of \code{JSON} data.
}
