% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_query.R
\name{show_query}
\alias{show_query}
\alias{show_query.tbl_lazy_json}
\title{Present computation steps.}
\usage{
\method{show_query}{tbl_lazy_json}(x, ...)
}
\arguments{
\item{x}{A \code{tbl_lazy_json} object.}

\item{...}{Filtering expressions.}
}
\description{
Present computation steps.
}
